+++ src/app/access/course/[courseId]/assignment/[assignmentId]/_components/reading_activity/reading_activity.tsx
@@ -1,13 +1,9 @@
 "use client";
 
-import dynamic from "next/dynamic";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { Course, Assignment, Activity, ActivityData, AssignmentData, ActivityType, Question } from '@prisma/client'
-import DynamicDocViewer from './dynamic-doc-viewer'
-import DocViewer, { DocViewerRenderers } from "@cyntler/react-doc-viewer";
 import { WebGazerManager } from '~/providers/WebGazerManager';
-
-import "./pdf-viewer.css";
+import PDFViewer from './pdf-viewer';
 
 type ReadingActivityProps = {
   course: Course
@@ -45,35 +41,11 @@ export default function ReadingActivity(props: ReadingActivityProps) {
     webGazer.end();
     webGazer = new WebGazerManager();
   };
-
-  const docs = [
-    { uri: "https://arxiv.org/pdf/1708.08021.pdf" }, // Remote file
-  ];
-
-  useEffect(() => {
-    const element = document.getElementById('pdf-controls');
-    if (element) {
-      element.remove();
-    }
-  }, []);
   
   return (
     <>
       <div className="w-full flex justify-center items-center">
-        <DocViewer
-          style={{ width: "70%", height: "70%" }}
-          documents={docs} 
-          pluginRenderers={DocViewerRenderers}
-          config={{
-            header: {
-              overrideComponent: () => <div>Custom Header</div>,
-              disableHeader: true,
-              disableFileName: true,
-              retainURLParams: true
-            },
-            pdfVerticalScrollByDefault: true
-          }} 
-        />
+        <PDFViewer/>
       </div>
     </>
   )
